devops/api_chat_bd/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.3.0
qdrant-client==1.7.0
langchain==0.1.9
langchain-community==0.0.27
langchain-google-genai==0.0.11
google-generativeai>=0.4.1,<0.5.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-dotenv==1.0.0
pypdf==4.0.1